[refactor] UIImage+resize 위치 변경
