+++ SpaceCapsule/SpaceCapsule/Extensions/UIImage+.swift
@@ -36,38 +36,3 @@ extension UIImage {
     static let refresh = UIImage(systemName: "arrow.clockwise")
 }
 
-extension UIImage {
-    func resize(_ newWidth: CGFloat) -> UIImage {
-        let scale = newWidth / self.size.width
-        let newHeight = self.size.height * scale
-
-        let size = CGSize(width: newWidth, height: newHeight)
-        let render = UIGraphicsImageRenderer(size: size)
-        let renderImage = render.image { _ in
-            self.draw(in: CGRect(origin: .zero, size: size))
-        }
-
-        return renderImage
-    }
-
-    static func resize(data: Data, to targetSize: CGSize, scale: CGFloat) -> UIImage? {
-        let imageSourceOptions = [kCGImageSourceShouldCache: false] as CFDictionary
-        guard let imageSource = CGImageSourceCreateWithData(data as CFData, imageSourceOptions) else {
-            return nil
-        }
-
-        let maxDimension = max(targetSize.width, targetSize.height) * scale
-        let resizingOptions = [
-            kCGImageSourceCreateThumbnailFromImageAlways: true,
-            kCGImageSourceShouldCacheImmediately: true,
-            kCGImageSourceCreateThumbnailWithTransform: true,
-            kCGImageSourceThumbnailMaxPixelSize: maxDimension,
-        ] as CFDictionary
-
-        guard let resizedImage = CGImageSourceCreateThumbnailAtIndex(imageSource, 0, resizingOptions) else {
-            return nil
-        }
-
-        return UIImage(cgImage: resizedImage)
-    }
-}
